backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
boto3>=1.34.129
requests-oauthlib>=2.0.0
cryptography>=42.0.8
python-dotenv>=1.0.1
motor>=3.3.2
pymongo>=4.6.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
ipfshttpclient>=0.8.0a2
web3>=6.15.1
Pillow>=10.2.0
tensorflow>=2.15.0
numpy>=1.24.3
aiofiles>=23.2.1
